Turn DotExpr props into strings when parsing
The dot notation for attribute access only allows a lookup by name.
Therefore, `DotExpr` can be seen as `expression * string` rather than
`expression * expression`.  It is still parsed as `expr DOT ident` but the
attribute name is turned into a string with `ident_name` in the parser.
This makes the handling of `DotExpr` values easier since there is no need to
match it with an `expression` pattern anymore.