Install .cmx files
Since OCaml 4.03, the compiler can use .cmx files to perform
cross-module optimization. A side effect of this is that if these files
are not installed and the corresponding .cmi file is not compiled with
`-opaque`, the following warning is printed:

```
Warning 58: no cmx file was found in path for module Jg_template, and
its interface was not compiled with -opaque
```

This fixes this warning and makes it possible to get extra
optimizations.