Raise errors when macro gets called with unknown arguments.
This will cause us to raise if
- a macro receives more anonymous arguments than it expects
- a macro receives a keyword argument that it does not expect

This behaviour is less surprising and is helpful for people that move
over from jinja2 since it means their templates do not silently
misbehave where they used to pass in positional arguments for
arguments that were defined with a default value.
e.g.
{% macro test(a=False, b=False) %} ... {% endmacro %}
{{ test(True, True) }}